chore: improve interfaces, add clarity, extract Builder
